Refactoring

+++ app/src/main/java/com/xpenses/android/feature_expenses/data/local/relation/MonthAndExpenditureRelation.kt
@@ -2,5 +2,5 @@ package com.xpenses.android.feature_expenses.data.local.relation
 
 data class MonthAndExpenditureRelation(
     val month: String,
-    val expenditure: Long
+    val expenditure: Double
 )
